+++ src/jsMain/kotlin/at/posselt/pfrpg2e/kingdom/KingdomFeature.kt
@@ -55,6 +55,7 @@ fun KingdomFeature.explodeLevels(): List<RawExplodedKingdomFeature> =
             skillIncrease = skillIncrease,
             kingdomFeat = kingdomFeat,
             claimHexAttempts = claimHexAttempts,
+            automationNotes = automationNotes,
         )
     }