+++ src/jsMain/kotlin/at/posselt/pfrpg2e/kingdom/dialogs/AddEvent.kt
@@ -208,6 +208,7 @@ class AddEvent(
                                 actor = kingdomActor,
                                 kingdom = kingdom,
                                 elem = elem,
+                                activityId = null,
                             )
                         }
                     }