+++ src/jsMain/kotlin/at/posselt/pfrpg2e/kingdom/sheet/contexts/GovernmentContext.kt
@@ -6,6 +6,7 @@ import at.posselt.pfrpg2e.app.forms.SelectOption
 import at.posselt.pfrpg2e.kingdom.RawFeat
 import at.posselt.pfrpg2e.kingdom.RawGovernment
 import at.posselt.pfrpg2e.kingdom.data.RawGovernmentChoices
+import at.posselt.pfrpg2e.kingdom.formatRequirements
 import at.posselt.pfrpg2e.toLabel
 import js.objects.JsPlainObject
 
@@ -20,6 +21,7 @@ external interface GovernmentContext {
     val boosts: String
     val featDescription: String
     val featAutomationNotes: String?
+    val featRequirements: String?
 }
 
 fun RawGovernmentChoices.toContext(
@@ -46,6 +48,7 @@ fun RawGovernmentChoices.toContext(
         feat = feat?.name ?: "",
         featDescription = feat?.text ?: "",
         featAutomationNotes = feat?.automationNotes,
+        featRequirements = feat?.requirements?.formatRequirements(feats),
         abilityBoosts = abilityBoosts.toContext(
             prefix = "government.",
             free = government?.freeBoosts ?: 0,